+++ tests/ManagerTest.php
@@ -92,19 +92,4 @@ class ManagerTest extends BaseTestCase
 		$newRepository = Manager::new();
 		$this->assertSame(22, $newRepository->config('port'));
 	}
-
-	private function checkStorageDir()
-	{
-		$dir = __DIR__ . '/../storage';
-		if(file_exists($dir)){
-			$it = new \RecursiveDirectoryIterator($dir, \FilesystemIterator::SKIP_DOTS);
-			$it = new \RecursiveIteratorIterator($it, \RecursiveIteratorIterator::CHILD_FIRST);
-			foreach($it as $file) {
-				if ($file->isDir()) rmdir($file->getPathname());
-				else unlink($file->getPathname());
-			}
-		}else{
-			mkdir($dir);
-		}
-	}
 }